Reorganize CharField related class so they can be used from other Djangae fields
